# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "@aasumitro",
            "url": "https://aasumitro.id/",
            "email": "hello@aasumitro.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/aasumitro/posbe/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/v1/store/prefs": {
            "get": {
                "description": "Get Store Settings List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Store Settings",
                "responses": {
                    "200": {
                        "description": "OK RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.StoreSetting"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "put": {
                "description": "Update Floor Data by ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store"
                ],
                "summary": "Update Floor Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "value",
                        "name": "value",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.StoreSetting"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.StoreSetting": {
            "type": "object",
            "additionalProperties": true
        },
...
        "utils.ErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.ValidationErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}